+++ pandas/core/interchange/_from_dataframe.py
