+++ pandas/tests/interchange/test_impl.py
@@ -14,7 +14,7 @@ from pandas.core.interchange.dataframe_protocol import (
     ColumnNullType,
     DtypeKind,
 )
-from pandas.core.interchange.from_dataframe import from_dataframe
+from pandas.core.interchange._from_dataframe import from_dataframe
 
 test_data_categorical = {
     "ordered": pd.Categorical(list("testdata") * 30, ordered=True),
